correct location of BaseSpecification class
