+++ docs/1-creatingSpecs.md
@@ -52,7 +52,7 @@ public function modify(QueryBuilder $qb, $dqlAlias)
 
 ## BaseSpecification
 
-To make your life easier you may use the `Happyr\DoctrineSpecification\BaseSpecification` class. When you extend
+To make your life easier you may use the `Happyr\DoctrineSpecification\Specification\BaseSpecification` class. When you extend
 this class you don't need to bother with `getFilter` or `modify`. You need to do 2 things:
 
 1. If you implement a constructor, make sure to call the parent constructor with $dqlAlias
@@ -61,7 +61,7 @@ this class you don't need to bother with `getFilter` or `modify`. You need to do
 Consider the following example.
 
 ```php
-use Happyr\DoctrineSpecification\BaseSpecification;
+use Happyr\DoctrineSpecification\Specification\BaseSpecification;
 use Happyr\DoctrineSpecification\Spec;
 
 /**
